+++ samples/hello/hello_menu_bar.rb
@@ -43,6 +43,11 @@ root { |r|
         message_box(parent: r, title: 'About', message: 'About my application.')
       end
     }
+    menu_item(:preferences) {
+      on('command') do
+        message_box(parent: r, title: 'Preferences', message: 'Preferences of my application.')
+      end
+    }
     menu(label: 'File', underline: 0) {
       menu_item(label: 'New', underline: 0) {
         accelerator 'Command+N'